Refac: "프로젝트 초대 이메일 템플릿 수정"
