+++ src/main/resources/templates/InviteEmail.html
@@ -8,12 +8,12 @@
 </head>
 <body>
     <div>
-        <p>안녕하세요. <span th:text="${nickname}"></span>님</p>
+        <h3>안녕하세요🌟 <span th:text="${nickname}"></span>님 개발일지 프로젝트 Prolog 입니다.</h3>
 
-        <h2 th:text="${message}">메시지</h2>
+        <p th:text="${message}">메시지</p>
 
         <div>
-            <a th:href="${link+data}" th:text="${linkName}">Link</a>
+            <a th:href="${link+data}">🔗LINK</a>
         </div>
     </div>
 </body>
