+++ src/test/java/com/prolog/prologbackend/Project/Service/ProjectServiceTest.java
@@ -32,7 +32,7 @@ import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-class ProjectServiceImplTest {
+class ProjectServiceTest {
 
     @Mock
     private ProjectRepository projectRepository;
@@ -52,7 +52,7 @@ class ProjectServiceImplTest {
 
 
     @InjectMocks
-    private ProjectServiceImpl projectService;
+    private ProjectService projectService;
 
     @Test
     @DisplayName("Get프로젝트")
